Free shit, no updates, as of March 2021.
Free shit, no updates, as of March 2021. 

With so many issues because of the new closed repo, so many people are so poor that they can't even donate, we decided to make free shit available in the old free repo for everyone.  

All m3us and also the latest m3u's with the daily updates are ONLY available in the NEW CLOSED REPO.